+++ client-common/src/main/java/org/realityforge/replicant/client/transport/DataLoaderService.java
@@ -8,10 +8,25 @@ import javax.annotation.Nullable;
  */
 public interface DataLoaderService<G>
 {
+  enum State
+  {
+    /// The service is not yet connected or has been disconnected
+    DISCONNECTED,
+    /// The service has started connecting but connection has not completed.
+    CONNECTING,
+    /// The service is connected.
+    CONNECTED,
+    /// The service has started disconnecting but disconnection has not completed.
+    DISCONNECTING,
+    /// The service is in error state. This error may occur during connection, disconnection or in normal operation.
+    ERROR
+  }
+
   /**
-   * @return true if connected to underlying data source.
+   * Return the state of the service.
    */
-  boolean isConnected();
+  @Nonnull
+  State getState();
 
   /**
    * Connect to the underlying data source.