+++ client-common/src/test/java/org/realityforge/replicant/client/transport/TestDataLoadService.java
@@ -78,12 +78,6 @@ final class TestDataLoadService
   }
 
   @Override
-  public boolean isConnected()
-  {
-    return true;
-  }
-
-  @Override
   protected void doConnect( @Nullable final Runnable runnable )
   {
   }